# contracts/runtime/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/sessions": {
      "get": {
        "operationId": "SessionController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List allowlisted Gateway sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{id}": {
      "get": {
        "operationId": "SessionController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read a Gateway session from the durable read model",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{id}/sync": {
      "post": {
        "operationId": "SessionController_requestSync",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Queue a full session, group and group-member sync",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{id}/sync-runs/{runId}": {
      "get": {
        "operationId": "SessionController_getSyncRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read sync progress",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/groups": {
      "get": {
        "operationId": "GroupController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "description": "Gateway session owning the read model",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List active groups from the Runtime read model",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}": {
      "get": {
        "operationId": "GroupController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "description": "Gateway session owning the group",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read a group and its synchronized members",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/members": {
      "get": {
        "operationId": "GroupController_members",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "description": "Gateway session owning the read model",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupMemberListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List synchronized group members without a contacts dependency",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/refresh-capability": {
      "post": {
        "operationId": "GroupController_refreshCapability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "description": "Gateway session owning the group",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accepted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Request an asynchronous capability refresh for one group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/v1/message-jobs": {
      "post": {
        "operationId": "MessageJobController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Create an idempotent scheduled text-message job",
        "tags": [
          "message-jobs"
        ]
      }
    },
    "/api/v1/message-jobs/{id}": {
      "get": {
        "operationId": "MessageJobController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read current durable message-job state",
        "tags": [
          "message-jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Automation Runtime API",
    "description": "Stable, versioned API contract consumed by WA Studio",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "runtime-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Runtime-Key"
      }
    },
    "schemas": {
      "SessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "ready"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "pushName": {
            "type": "string",
            "nullable": true
          },
          "connectedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastActiveAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "engineLoaded": {
            "type": "boolean"
          },
          "lastError": {
            "type": "string",
            "nullable": true
          },
          "restriction": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "gatewayCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "gatewayUpdatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "syncedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "phone",
          "pushName",
          "connectedAt",
          "lastActiveAt",
          "engineLoaded",
          "lastError",
          "restriction",
          "gatewayCreatedAt",
          "gatewayUpdatedAt",
          "syncedAt"
        ]
      },
      "SessionListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "SyncRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "syncType": {
            "type": "string",
            "example": "FULL"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "groupsSynced": {
            "type": "number"
          },
          "membersSynced": {
            "type": "number"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sessionId",
          "syncType",
          "status",
          "groupsSynced",
          "membersSynced",
          "error",
          "requestedAt",
          "startedAt",
          "completedAt"
        ]
      },
      "GroupSendCapabilityDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ALLOWED",
              "DENIED",
              "UNKNOWN"
            ]
          },
          "reason": {
            "type": "string"
          },
          "checkedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "invalidatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "status",
          "reason",
          "checkedAt",
          "invalidatedAt",
          "revision"
        ]
      },
      "GroupDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "example": "120363000000000000@g.us"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "linkedParentId": {
            "type": "string",
            "nullable": true
          },
          "participantsCount": {
            "type": "number",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean",
            "nullable": true
          },
          "isReadOnly": {
            "type": "boolean",
            "nullable": true
          },
          "isAnnounce": {
            "type": "boolean",
            "nullable": true
          },
          "settingsLocked": {
            "type": "boolean",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "detailsSyncedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "syncedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sendCapability": {
            "$ref": "#/components/schemas/GroupSendCapabilityDto"
          }
        },
        "required": [
          "sessionId",
          "id",
          "name",
          "description",
          "ownerId",
          "linkedParentId",
          "participantsCount",
          "isAdmin",
          "isReadOnly",
          "isAnnounce",
          "settingsLocked",
          "isActive",
          "detailsSyncedAt",
          "syncedAt",
          "sendCapability"
        ]
      },
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "limit",
          "offset"
        ]
      },
      "GroupListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "GroupMemberDto": {
        "type": "object",
        "properties": {
          "participantId": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "isSuperAdmin": {
            "type": "boolean"
          }
        },
        "required": [
          "participantId",
          "phoneNumber",
          "displayName",
          "isAdmin",
          "isSuperAdmin"
        ]
      },
      "GroupDetailDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "example": "120363000000000000@g.us"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "linkedParentId": {
            "type": "string",
            "nullable": true
          },
          "participantsCount": {
            "type": "number",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean",
            "nullable": true
          },
          "isReadOnly": {
            "type": "boolean",
            "nullable": true
          },
          "isAnnounce": {
            "type": "boolean",
            "nullable": true
          },
          "settingsLocked": {
            "type": "boolean",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "detailsSyncedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "syncedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sendCapability": {
            "$ref": "#/components/schemas/GroupSendCapabilityDto"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberDto"
            }
          }
        },
        "required": [
          "sessionId",
          "id",
          "name",
          "description",
          "ownerId",
          "linkedParentId",
          "participantsCount",
          "isAdmin",
          "isReadOnly",
          "isAnnounce",
          "settingsLocked",
          "isActive",
          "detailsSyncedAt",
          "syncedAt",
          "sendCapability",
          "members"
        ]
      },
      "GroupMemberListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CreateMessageJobDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "example": "c4984920-cc62-4720-97ca-c46b4f3f9d80",
            "description": "Must be in the Runtime session allowlist"
          },
          "recipientId": {
            "type": "string",
            "example": "120363000000000000@g.us"
          },
          "text": {
            "type": "string",
            "example": "Hello from Automation Runtime",
            "maxLength": 4096
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-08-11T12:00:00.000Z",
            "description": "Defaults to now"
          },
          "dryRun": {
            "type": "boolean",
            "default": true,
            "description": "Live sends also require ALLOW_LIVE_SENDS=true"
          }
        },
        "required": [
          "sessionId",
          "recipientId",
          "text"
        ]
      },
...
    }
  }
}